# api.yaml
openapi: 3.0.0
info:
  title: Proyecto WEB API
  version: 1.0.0
  description: Documentación de la API de Proyecto WEB
paths:
...
  /api/admin/products:
    post:
      summary: Agrega un producto
      description: Agrega un producto a la tienda; solo admins pueden usarlo.
      operationId: addNewProduct
      parameters:
        - name: token
          in: cookie
          required: true
          schema:
            type: string
            example: "edrfvm890234i0szadjflkln34tun9"
      requestBody:
        description: Datos del producto
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                stock:
                  type: integer
                  example: 10
                name:
                  type: string
                  example: "Vamos ekipocaraio"
                price:
                  type: integer
                  example: 123456
                image:
                  type: string
                  example: "https://pbs.twimg.com/media/GABUcWXQAAAtV6.jpg"
                description:
                  type: string
                  example: "Un producto de calidad"
      responses:
        '200':
          description: Producto creado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
        '401':
          description: El usuario no se encuentra autenticado
        '403':
          description: El usuario no tiene permisos para acceder a este endpoint
  /api/admin/products/{id}:
    delete:
      summary: Elimina un producto
      description: Elimina un producto de la tienda según su id; solo los administradores pueden usar esta operación.
      operationId: deleteProduct
      parameters:
        - name: token
          in: cookie
          required: true
          schema:
            type: string
            example: "edrfvm890234i0szadjflkln34tun9"
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Producto eliminado exitosamente
        '401':
          description: El usuario no se encuentra autenticado
        '403':
          description: El usuario no tiene permisos para acceder a este endpoint
        '404':
          description: El producto no existe en la base de datos
  /api/admin/product/{id}:
    post:
      summary: Modifica un producto
      description: Modifica un producto de la tienda según su id; solo los administradores pueden usar esta operación.
      operationId: updateProduct
      parameters:
        - name: token
          in: cookie
          required: true
          schema:
            type: string
            example: "edrfvm890234i0szadjflkln34tun9"
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        description: Datos del producto
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                stock:
                  type: integer
                  example: 10
                name:
                  type: string
                  example: "Vamos ekipo 2.1"
                price:
                  type: integer
                  example: 1234
                image:
                  type: string
                  example: "https://pbs.twimg.com/media/GABUcWXQAAtZV6.jpg"
                description:
                  type: string
                  example: "Un producto de calidad"
      responses:
        '200':
          description: Producto modificado exitosamente
        '401':
          description: El usuario no se encuentra autenticado
        '403':
          description: El usuario no tiene permisos para acceder a este endpoint
        '404':
          description: El producto no existe en la base de datos
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: Verificación cookie-token
security:
  - cookieAuth: []